--- egg/zoo/emcom_as_ssl/scripts/o_test_evaluation.py
@@ -1,4 +1,0 @@
-# Copyright (c) Facebook, Inc. and its affiliates.
-
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
